LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/worker-src-none/sharedworker-import-data.https-expected.txt


Harness Error (FAIL), message = Unhandled rejection: null is not an object (evaluating 'resolveObject.addEventListener')

FAIL Content Security Policy: Expects blocked for sharedworker-import-data to cross-https origin and keep-origin redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to cross-https origin and no-redirect redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to cross-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to same-https origin and keep-origin redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to same-https origin and keep-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to same-https origin and no-redirect redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to same-https origin and no-redirect redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to same-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-import-data to same-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
